fully extracted configured addon
