+++ addon_service/configured_storage_addon/views.py
@@ -4,40 +4,16 @@ from rest_framework.decorators import action
 from rest_framework.response import Response
 
 from addon_service.common.credentials_formats import CredentialsFormats
-from addon_service.common.permissions import (
-    IsAuthenticated,
-    IsValidHMACSignedRequest,
-    SessionUserCanViewReferencedResource,
-    SessionUserIsOwner,
-    SessionUserMayConnectAddon,
-)
-from addon_service.common.viewsets import RetrieveWriteDeleteViewSet
 from addon_service.common.waterbutler_compat import WaterButlerCredentialsSerializer
 
+from ..abstract.configured_addon.views import ConfiguredAddonViewSet
 from .models import ConfiguredStorageAddon
 from .serializers import ConfiguredStorageAddonSerializer
 
 
-class ConfiguredStorageAddonViewSet(RetrieveWriteDeleteViewSet):
+class ConfiguredStorageAddonViewSet(ConfiguredAddonViewSet):
     queryset = ConfiguredStorageAddon.objects.active()
     serializer_class = ConfiguredStorageAddonSerializer
-
-    def get_permissions(self):
-        match self.action:
-            case "retrieve" | "retrieve_related":
-                return [IsAuthenticated(), SessionUserCanViewReferencedResource()]
-            case "partial_update" | "update" | "destroy":
-                return [IsAuthenticated(), SessionUserIsOwner()]
-            case "create":
-                return [IsAuthenticated(), SessionUserMayConnectAddon()]
-            case "get_wb_credentials":
-                return [IsValidHMACSignedRequest()]
-            case None:
-                return super().get_permissions()
-            case _:
-                raise NotImplementedError(
-                    f"no permission implemented for action '{self.action}'"
-                )
 
     @action(
         detail=True,
